refactor(sidecar): split proxy.py into the klangksidecar package (#2450)
Breaks the monolithic src/klangksidecar/proxy.py (~1938 lines) into a
multi-module package: config (env), state (shared mutable containers),
allowlist (FQDN spec matching + in-session consent memory), rules (iptables
rule + learned-IP lifecycle), resolve (DNS wire helpers + per-query server),
packets (IPv4 parse + forged-RST), consent (the egress-sidecar WS client),
nfqueue (the SYN consent gate), and app (the asyncio loop + main). __init__
re-exports every name so the flat ``klangksidecar.X`` API the tests + callers
use keeps working; __main__ makes ``python -m klangksidecar`` the PID-1 entry
(entrypoint.sh updated).

The split preserves the single-namespace test semantics: reads of
``klangksidecar.X`` resolve via the re-export; singleton stdlib modules
(time/subprocess/socket/dns) are patched globally; and the ~100
monkeypatch.setattr calls retarget to the defining submodule
(proxy.rules.allow, proxy.resolve.query_name, ...). The two reassigned
session lists (_SESSION_HOST_ALLOWS/_DENIES) are read qualified (state.X) so
a test reassign is visible to the code; all other state is mutated in place
(shared by import).

No behavior change: 148 unit tests + the 10 real-podman e2e tests pass
unchanged, and the sidecar image still installs the wheel + runs
``python3 -m klangksidecar``. src/containers/network/proxy.py is untouched
(still unit-tested by the klangk suite); the package is what ships now.
