+++ src/klangksidecar/klangksidecar/allowlist.py
@@ -1,0 +1,307 @@
+"""FQDN allow/deny-list matching + in-session consent memory (#2450, #2377).
+
+Static specs (parse_specs from KLANGKNETWORK_EGRESS_ALLOW / _REJECT) plus the
+dynamic per-session host allow/deny lists a consent verdict installs
+(#2372/#2434/#2446). ports_for / rejected_for are the DNS gate; the
+_session_host_*_ttl helpers are the NFQUEUE gate's last-chance short-circuit.
+"""
+
+from __future__ import annotations
+
+import os
+import time
+
+from . import (
+    state,
+)  # _SESSION_HOST_ALLOWS/_DENIES read qualified so tests can reassign them
+
+# Host-scope modes for an allow-list spec, nginx-style (#2377): a bare host is
+# EXACT (apex only); a leading-dot ``.host`` is INCLUSIVE (apex + subdomains);
+# ``*.host`` is SUBDOMAINS only. (Bare = exact is the breaking flip from the
+# old "bare = apex+subdomains" model.) One definition shared by parse_specs /
+# ports_for / _session_host_allows_ttl.
+_EXACT = "exact"
+
+
+_INCLUSIVE = "inclusive"
+
+
+_SUBDOMAINS = "subdomains"
+
+
+def parse_specs(
+    env_var: str = "KLANGKNETWORK_EGRESS_ALLOW",
+) -> list[tuple[str, int | None, str]]:
+    """Structured host specs from ``env_var`` (#2377, #2367).
+
+    Each entry is ``(host, port, mode)``: ``mode`` is :data:`_EXACT` (bare host,
+    apex only), :data:`_INCLUSIVE` (``.host``, apex + subdomains), or
+    :data:`_SUBDOMAINS` (``*.host``, subdomains only). ``port`` is ``None`` for
+    all-ports. CIDR specs (``10.0.0.0/8``) are excluded — the entrypoint applies
+    those statically. The grammar mirrors ``klangk.netfilter.parse_allowed_domains``.
+    """
+    out: list[tuple[str, int | None, str]] = []
+    for spec in os.environ.get(env_var, "").split(","):
+        spec = spec.strip()
+        if not spec or "/" in spec:
+            continue
+        port: int | None = None
+        if ":" in spec:
+            host_part, port_part = spec.rsplit(":", 1)
+            if port_part.isdigit():
+                port = int(port_part)
+                spec = host_part
+        s = spec.lower()
+        mode = _EXACT
+        if s.startswith("*."):
+            mode = _SUBDOMAINS
+            s = s[2:]
+        elif s.startswith("."):
+            mode = _INCLUSIVE
+            s = s[1:]
+        if s:
+            out.append((s, port, mode))
+    return out
+
+
+SPECS = parse_specs()
+
+
+# Static deny-list specs from ``KLANGKNETWORK_EGRESS_REJECT`` (#2367): a name
+# matching one of these is NXDOMAIN'd unconditionally (see :func:`rejected_for`).
+REJECT_SPECS = parse_specs("KLANGKNETWORK_EGRESS_REJECT")
+
+
+def _host_matches(qname: str, host: str, mode: str) -> bool:
+    """Does ``qname`` match ``host`` under nginx-style scope ``mode`` (#2377)?
+
+    Shared by :func:`ports_for` (the DNS gate) and :func:`_session_host_allows_ttl`
+    (the NFQUEUE gate) so the two can't drift. :data:`_EXACT` (bare host) matches
+    the apex only; :data:`_INCLUSIVE` (``.host``) matches apex + subdomains;
+    :data:`_SUBDOMAINS` (``*.host``) matches subdomains only. The suffix check
+    requires a leading dot, so ``evilexample.com`` does NOT match
+    ``example.com``.
+    """
+    if mode == _SUBDOMAINS:
+        return qname.endswith("." + host)
+    if mode == _INCLUSIVE:
+        return qname == host or qname.endswith("." + host)
+    return qname == host  # _EXACT (and the safe default for an unknown mode)
+
+
+def ports_for(qname: str) -> set[int] | None:
+    """The ports a queried name is allowed on under :data:`SPECS` (#2377).
+
+    ``None``  — a port-less spec matched (allow all ports).
+    ``set()`` — nothing matched (deny).
+    ``{443, ...}`` — allow exactly these TCP ports.
+
+    Scope: a bare host is :data:`_EXACT` (apex only); ``.host`` is
+    :data:`_INCLUSIVE` (apex + subdomains); ``*.host`` is :data:`_SUBDOMAINS`
+    (subdomains only, apex excluded).
+    """
+    ports: set[int] = set()
+    _prune_session_allows()
+    session = [(h, p, m) for (h, p, m, _exp) in state._SESSION_HOST_ALLOWS]
+    for host, port, mode in (*SPECS, *session):
+        if not _host_matches(qname, host, mode):
+            continue
+        if port is None:
+            return None  # an all-ports spec dominates
+        ports.add(port)
+    return ports
+
+
+def rejected_for(qname: str) -> bool:
+    """Does ``qname`` match a :data:`REJECT_SPECS` entry (#2367)?
+
+    Parallel to :func:`ports_for` for the deny-list, but boolean -- a rejected
+    name is NXDOMAIN'd unconditionally, so there is no port dimension. Matches
+    via :func:`_host_matches` (nginx-style scope): bare = apex only, ``.host`` =
+    apex + subdomains, ``*.host`` = subdomains only.
+    """
+    return any(_host_matches(qname, host, mode) for host, _port, mode in REJECT_SPECS)
+
+
+def _prune_session_allows() -> None:
+    """Drop expired in-session host allows (lazy sweep, #2434).
+
+    :data:`_SESSION_HOST_ALLOWS` is loop-only (no lock), so -- unlike
+    :data:`_LEARNED` / :data:`_REJECTED`, which are swept off-loop by
+    :func:`sweep_once` under :data:`_LOCK` -- its timed entries expire here, on
+    the loop, the next time a gate (:func:`ports_for`,
+    :func:`_session_host_allows_ttl`, :func:`_add_session_host`) reads them.
+    Cheap (the list is tiny -- one entry per consented host:port) and keeps the
+    structure from growing unbounded across a long session.
+    """
+    now = time.time()
+    state._SESSION_HOST_ALLOWS[:] = [
+        t for t in state._SESSION_HOST_ALLOWS if t[3] > now
+    ]
+
+
+def _add_session_host(host: str, port: int, ttl: float) -> None:
+    """Allow-list ``host:port`` in-session for a consent allow verdict (#2372,
+    #2434).
+
+    Adds ``(host, port, _EXACT, now + ttl)`` to :data:`_SESSION_HOST_ALLOWS` so
+    :func:`ports_for` (the DNS gate) treats the host as allow-listed for the
+    verdict's lifetime -- the DNS path then learns every resolved IP and allows
+    it without NFQUEUE, so a CDN-rotated IP no longer re-prompts (or, if it
+    still races NFQUEUE, :func:`_session_host_allows_ttl` short-circuits it in
+    :func:`_cb`). EXACT scope: the user approved the specific qname they saw, so
+    only that host (not its subdomains) is opened (#2377). Deduped; a re-allow
+    of the same host:port refreshes the expiry (``max`` -- never shortens an
+    unexpired entry). A timed allow (5s/5m/1h/tilrestart) is host-scoped just
+    like ``forever`` (#2434); ``once`` carries no host-allow (per-connection, so
+    a reconnect re-prompts). Loop-only (no lock).
+    """
+    _prune_session_allows()
+    expire = time.time() + ttl
+    spec = (host, port, _EXACT)
+    for i, (h, p, mode, _exp) in enumerate(state._SESSION_HOST_ALLOWS):
+        if (h, p, mode) == spec:
+            state._SESSION_HOST_ALLOWS[i] = (h, p, mode, max(_exp, expire))
+            return
+    state._SESSION_HOST_ALLOWS.append((host, port, _EXACT, expire))
+
+
+def _session_host_allows_ttl(host: str, port: int) -> float | None:
+    """Remaining seconds an in-session allow covers ``host`` on ``port``, or
+    ``None`` (#2372, #2434).
+
+    Used by :func:`_cb` as the last-chance gate before prompting: a SYN to a
+    host:port the user allowed (timed or forever) -- including a CDN-rotated or
+    resolver-cached IP that no fresh DNS resolution re-ACCEPTed -- is
+    auto-allowed, learned for the allow's remaining window, so the user isn't
+    re-asked (and a hold timeout can't fail-close a still-allowed host to a
+    deny, #2434). Matches via :func:`_host_matches` (nginx-style scope); entries
+    are added EXACT by :func:`_add_session_host`, so only the approved host
+    matches (#2377). Returns the max remaining TTL across matching entries.
+    Loop-only (no lock).
+    """
+    if not host:
+        return None
+    _prune_session_allows()
+    now = time.time()
+    best: float | None = None
+    for h, p, mode, exp in state._SESSION_HOST_ALLOWS:
+        if exp <= now:
+            continue  # belt-and-suspenders: _prune ran above, but a just-expired
+            # entry can survive the microseconds between its `now` and this one.
+        if _host_matches(host, h, mode) and (p == port or p is None):
+            remaining = exp - now
+            if best is None or remaining > best:
+                best = remaining
+    return best
+
+
+def _prune_session_denies() -> None:
+    """Drop expired in-session host denies (lazy sweep, #2446).
+
+    :data:`_SESSION_HOST_DENIES` is loop-only (no lock), so -- like
+    :data:`_SESSION_HOST_ALLOWS` -- its timed entries expire here, on the loop,
+    the next time a gate (:func:`_session_host_denies_ttl`,
+    :func:`_add_session_deny`, :func:`_drop_session_denies`) reads them. Cheap
+    (the list is tiny -- one entry per denied host:port) and keeps the structure
+    from growing unbounded across a long session.
+    """
+    now = time.time()
+    state._SESSION_HOST_DENIES[:] = [
+        t for t in state._SESSION_HOST_DENIES if t[3] > now
+    ]
+
+
+def _add_session_deny(host: str, port: int, ttl: float) -> None:
+    """Deny ``host:port`` in-session for a consent deny verdict (#2446).
+
+    The deny-side mirror of :func:`_add_session_host`: adds
+    ``(host, port, _EXACT, now + ttl)`` to :data:`_SESSION_HOST_DENIES` so
+    :func:`_cb` (via :func:`_session_host_denies_ttl`) suppresses a re-prompt
+    for a host the user already denied -- including a CDN-rotated or
+    resolver-cached IP that the per-IP :data:`_REJECTED` rule does not cover
+    (the CARRYOVER-SURPRISE, #2446). EXACT scope (only the denied host, not its
+    subdomains, #2377); deduped, a re-deny refreshes the expiry (``max`` --
+    never shortens an unexpired entry). ``once`` adds nothing (per-connection,
+    so a reconnect re-prompts). Loop-only (no lock).
+    """
+    _prune_session_denies()
+    expire = time.time() + ttl
+    spec = (host, port, _EXACT)
+    for i, (h, p, mode, _exp) in enumerate(state._SESSION_HOST_DENIES):
+        if (h, p, mode) == spec:
+            state._SESSION_HOST_DENIES[i] = (h, p, mode, max(_exp, expire))
+            return
+    state._SESSION_HOST_DENIES.append((host, port, _EXACT, expire))
+
+
+def _session_host_denies_ttl(host: str, port: int) -> float | None:
+    """Remaining seconds an in-session deny covers ``host`` on ``port``, or
+    ``None`` (#2446).
+
+    The deny-side mirror of :func:`_session_host_allows_ttl`, used by
+    :func:`_cb` as the last-chance gate before prompting: a SYN to a host:port
+    the user already denied (timed or forever) -- including a CDN-rotated or
+    resolver-cached IP that no fresh per-IP :data:`_REJECTED` rule covers -- is
+    denied fast (RST + short REJECT) without re-prompting. Matches via
+    :func:`_host_matches` (entries are added EXACT, so only the denied host
+    matches, #2377); port must match (or the entry is all-ports). Returns the
+    max remaining TTL across matching entries. Loop-only (no lock).
+    """
+    if not host:
+        return None
+    _prune_session_denies()
+    now = time.time()
+    best: float | None = None
+    for h, p, mode, exp in state._SESSION_HOST_DENIES:
+        if exp <= now:
+            continue  # belt-and-suspenders: _prune ran above, but a just-expired
+            # entry can survive the microseconds between its `now` and this one.
+        if _host_matches(host, h, mode) and (p == port or p is None):
+            remaining = exp - now
+            if best is None or remaining > best:
+                best = remaining
+    return best
+
+
+def _drop_session_hosts(host: str) -> None:
+    """Remove a host's in-session allow coverage (#2370, #2372, #2434).
+
+    Drops every :data:`_SESSION_HOST_ALLOWS` entry whose host matches
+    (case-insensitive). Called on the **event loop** by
+    :meth:`SidecarConsentClient._handle_drop_rule` **before** :func:`drop_for_host`
+    forks iptables in the executor: while that fork runs (~tens of ms), the
+    NFQUEUE consumer (:func:`_cb` -> :func:`_session_host_allows_ttl`) and the DNS
+    path (:func:`ports_for`) both read :data:`_SESSION_HOST_ALLOWS`, and a
+    SYN/resolve arriving in that window would otherwise re-install a fresh
+    ACCEPT (the host's remaining allow TTL, via :func:`allow`) that the revoke
+    never clears. Clearing it first makes both gates deny during the window, so
+    no fresh rule can be installed; :func:`drop_for_host` then removes the
+    existing ACCEPTs. :data:`_SESSION_HOST_ALLOWS` is loop-only (no lock) --
+    touched on the loop, never inside :func:`drop_for_host` (executor thread). A
+    deny revoke does not call this (a deny never adds to
+    :data:`_SESSION_HOST_ALLOWS`).
+    """
+    hl = host.lower()
+    state._SESSION_HOST_ALLOWS[:] = [
+        t for t in state._SESSION_HOST_ALLOWS if t[0].lower() != hl
+    ]
+
+
+def _drop_session_denies(host: str) -> None:
+    """Remove a host's in-session deny coverage (#2446).
+
+    The deny-side mirror of :func:`_drop_session_hosts`, called on the event
+    loop by :meth:`SidecarConsentClient._handle_drop_rule` for a ``denied``
+    revoke BEFORE :func:`drop_for_host` forks iptables in the executor: while
+    that fork runs (~tens of ms), :func:`_cb` reads :data:`_SESSION_HOST_DENIES`,
+    and a SYN arriving in that window would otherwise keep auto-denying (and
+    re-installing a REJECT for) the host the operator just un-denied. Clearing
+    it first lets the host re-prompt. :data:`_SESSION_HOST_DENIES` is loop-only
+    (no lock) -- touched on the loop, never inside :func:`drop_for_host`
+    (executor thread).
+    """
+    hl = host.lower()
+    state._SESSION_HOST_DENIES[:] = [
+        t for t in state._SESSION_HOST_DENIES if t[0].lower() != hl
+    ]
