+++ src/klangksidecar/klangksidecar/nfqueue.py
@@ -1,0 +1,291 @@
+"""NFQUEUE consumer: consent-gate the connection SYN pending a verdict (#2324, #2329, #2450).
+
+_setup_nfq_consumer binds the queue + drives it from the event loop; _cb
+classifies one queued SYN (cached-verdict fast path / in-session host gates /
+hand-off to _decide_and_verdict); _decide_and_verdict awaits the verdict and
+applies it. netfilterqueue is imported lazily.
+"""
+
+from __future__ import annotations
+
+import asyncio
+import time
+from typing import TYPE_CHECKING
+
+from . import packets, rules
+from .allowlist import (
+    _add_session_deny,
+    _add_session_host,
+    _session_host_allows_ttl,
+    _session_host_denies_ttl,
+)
+from .config import CONSENT_REJECT_TTL, QUEUE_NUM, VERDICT_CACHE_TTL, _duration_ttl
+from .packets import parse_dest, parse_syn_tuple
+from .rules import _host_for
+from .state import _BG_TASKS, _INFLIGHT, _VERDICT_CACHE
+
+if TYPE_CHECKING:
+    # ``SidecarConsentClient`` appears only in annotations (passed in, never
+    # constructed here); lazy import avoids a cycle with .consent.
+    from .consent import SidecarConsentClient  # noqa: allow-deferred-import (annotation-only; avoids a cycle)
+
+
+def _setup_nfq_consumer(client: SidecarConsentClient | None):
+    """Bind the sidecar's NFQUEUE + drive it from the event loop (#2324, #2329).
+
+    Consent gates the connection SYN, not the DNS query: a non-allow-listed name
+    resolves (the workspace gets the IP) and the first packet to that IP is
+    queued here pending a verdict -- so the human decision window is the kernel's
+    connect timeout (tcp_syn_retries ~= 127s), not the resolver's <=30s
+    getaddrinfo cap.
+
+    The queue is read on the event-loop thread via ``get_fd()`` + ``add_reader``
+    (netfilterqueue is otherwise synchronous). The per-packet callback
+    (:func:`_cb`) is **non-blocking**: it retains the packet + hands the verdict
+    wait to a task (:func:`_decide_and_verdict`), so a slow verdict on one flow
+    does NOT serialize others -- distinct flows are held concurrently
+    (netfilterqueue supports deferred verdicts; outstanding packets count
+    against the kernel queue size, and the iptables rate-limit bounds arrivals).
+    netfilterqueue is a sidecar-only dep, imported lazily so the module loads
+    without it. Returns the bound ``NetfilterQueue`` (so :func:`_shutdown` can
+    unbind it on SIGTERM, #2400) or ``None`` on failure.
+    """
+    try:
+        from netfilterqueue import NetfilterQueue  # noqa: allow-deferred-import (sidecar-only; lazy so the module loads without it)
+    except Exception as exc:
+        print(f"nfqueue: netfilterqueue unavailable ({exc})", flush=True)
+        return None
+    try:
+        nfq = NetfilterQueue()
+        nfq.bind(QUEUE_NUM, lambda pkt: _cb(pkt, client))
+        loop = asyncio.get_running_loop()
+        # When the netlink socket is readable, process all pending packets on
+        # this (loop) thread; _cb then hands each off to a verdict task.
+        loop.add_reader(nfq.get_fd(), lambda: _drain(nfq))
+        print(f"nfqueue consumer bound to queue {QUEUE_NUM}", flush=True)
+        return nfq
+    except Exception as exc:
+        print(f"nfqueue consumer failed: {exc}", flush=True)
+        return None
+
+
+def _drain(nfq) -> None:
+    """Process pending NFQUEUE messages (called when the socket is readable)."""
+    try:
+        nfq.run(block=False)
+    except Exception:
+        pass  # a transient read failure defers to the next readability event
+
+
+def _cb(pkt, client: SidecarConsentClient | None) -> None:
+    """Classify + route one queued SYN -- non-blocking (#2324, #2329).
+
+    A retransmit of an already-decided connection reuses the cached verdict
+    inline (fast path). A new connection is retained + handed to
+    :func:`_decide_and_verdict` so the verdict wait doesn't block the queue's
+    drain (distinct connections are held concurrently, not serialized behind
+    the first).
+    """
+    payload = pkt.get_payload()
+    # Connection-level key (src IP+port + dst IP+port): a SYN retransmit shares
+    # its connection's tuple, a NEW connection has a new source port. Keying the
+    # verdict cache + in-flight set on the connection -- not just (dst, port) --
+    # is what makes duration=once re-prompt every subsequent connection instead
+    # of reusing a prior allow for the same destination (#2361).
+    src_ip, src_port, tdst, tport, _ = parse_syn_tuple(payload)
+    if tport:  # TCP: full connection tuple
+        dst, port = tdst, tport
+    else:  # non-TCP (UDP/other): no source port -> destination granularity
+        dst, port = parse_dest(payload)
+    if not dst or client is None:
+        pkt.drop()  # unparseable / pure-static (no consent configured) -> drop
+        return
+    if not client.connected:
+        # Consent WS down: consent is unavailable, so fail the off-list SYN
+        # FAST (ECONNREFUSED) like a deny verdict -- NOT a bare drop. A bare
+        # drop makes the kernel retransmit the SYN for ~127s (tcp_syn_retries),
+        # dangling the connection (#2308: no consent available -> a clean,
+        # prompt denial, not a hang). Forge the eager-deny RST inline
+        # (non-blocking sendto) so THIS connect() gets ECONNREFUSED at once,
+        # + a short REJECT (tcp-reset) backstop off-loop so retransmits are
+        # RST'd above NFQUEUE, then drop. On-list egress is unaffected
+        # (learned ACCEPT rules sit above NFQUEUE); once the WS reconnects,
+        # fresh off-list egress prompts again. The REJECT TTL is short, so no
+        # rule lingers past the outage (#2413).
+        if port:
+            packets._send_rst(payload)
+            asyncio.get_running_loop().run_in_executor(
+                None, rules.reject, dst, port, CONSENT_REJECT_TTL
+            )
+        pkt.drop()
+        return
+    flow = (src_ip, src_port, dst, port)
+    now = time.time()
+    # SYN retransmit of an already-decided connection -> reuse the verdict so
+    # the kernel's retransmits (tcp_syn_retries) don't each re-prompt.
+    cached = _VERDICT_CACHE.get(flow)
+    if cached is not None and cached[1] > now:
+        if cached[0] == "allow":
+            pkt.accept()
+        else:
+            # Forge the eager-deny RST (a retried connect() to a denied flow
+            # fails fast too), then drop. TCP only (port 0 is non-TCP).
+            if port:
+                packets._send_rst(payload)
+            pkt.drop()
+        return
+    # A SYN retransmit that arrives WHILE this connection is still held (no
+    # verdict yet) must not spawn a second consent request: the in-flight task
+    # resolves it, and a request per retransmit piles up duplicates that linger
+    # past the first's resolution (#2345 e2e flake). Drop it -- the kernel sends
+    # another retransmit once the verdict lands, and that one hits the cache.
+    if flow in _INFLIGHT:
+        pkt.drop()
+        return
+    host = _host_for(dst)  # DNS name if resolved here, else the IP
+    # An in-session host allow covers the whole domain, timed or forever
+    # (#2372, #2434): a SYN to a host:port the user already allowed -- including
+    # a CDN-rotated or resolver-cached IP that no fresh DNS resolution
+    # re-ACCEPTed -- is auto-allowed here, the last gate before prompting, so
+    # the user isn't re-asked for a domain they allowed. This is the fix for the
+    # ALLOW-REFUSED mismatch (#2434): without it a timed allow only covered the
+    # IP resolved at decision time, so a CDN-rotated IP re-entered NFQUEUE and a
+    # hold timeout there fail-closed to a deny REJECT an in-effect allow should
+    # override.
+    remaining = _session_host_allows_ttl(host, port) if port else None
+    if remaining is not None:
+        # allow() forks iptables under _LOCK -- run it off the loop thread (the
+        # file's invariant; every other allow/learn/reject call is in the
+        # executor). Fire-and-forget: conntrack ESTABLISHED,RELATED carries THIS
+        # connection after pkt.accept(); the ACCEPT rule only helps FUTURE
+        # connections, and the _VERDICT_CACHE write below covers retransmits.
+        # Learn for the allow's remaining window (timed) or ~forever; port-scoped
+        # (the consented port) -- deliberately stricter than the consent-allow
+        # path's all-ports learn (allow(dst, None, ...)).
+        asyncio.get_running_loop().run_in_executor(
+            None, rules.allow, dst, port, remaining
+        )
+        pkt.accept()
+        _VERDICT_CACHE[flow] = ("allow", now + VERDICT_CACHE_TTL)
+        # NOTE (#2370): revoking an allow must also drop this host from
+        # _SESSION_HOST_ALLOWS and clear _VERDICT_CACHE, or a revoked allow keeps
+        # passing (ports_for re-allow-lists it + this cache reuses the verdict).
+        # Wired with the revoke path in #2370.
+        return
+    # An in-session host deny covers the whole domain, timed or forever
+    # (#2446): a SYN to a host:port the user already denied -- including a
+    # CDN-rotated or resolver-cached IP that the per-IP _REJECTED rule does not
+    # cover -- is denied fast here, before prompting, so the user isn't re-asked
+    # for a domain they denied (the CARRYOVER-SURPRISE). Checked AFTER the allow
+    # gate above so an in-effect allow still overrides an in-effect deny.
+    deny_remaining = _session_host_denies_ttl(host, port) if port else None
+    if deny_remaining is not None:
+        # Forge the eager-deny RST so connect() fails fast (ECONNREFUSED) at
+        # once; a REJECT for the deny's remaining window backstops retransmits
+        # off-loop (reject() forks iptables under _LOCK). TCP only (port 0 is
+        # non-TCP). The _VERDICT_CACHE write covers retransmits of THIS flow.
+        if port:
+            try:
+                packets._send_rst(payload)
+            except Exception:
+                pass
+            asyncio.get_running_loop().run_in_executor(
+                None, rules.reject, dst, port, deny_remaining
+            )
+        pkt.drop()
+        _VERDICT_CACHE[flow] = ("deny", now + VERDICT_CACHE_TTL)
+        return
+    pkt.retain()  # keep the payload valid past this callback (deferred verdict)
+    _INFLIGHT.add(flow)
+    t = asyncio.create_task(_decide_and_verdict(pkt, flow, dst, port, host, client))
+    _BG_TASKS.add(t)  # strong ref so the verdict task isn't GC'd
+    t.add_done_callback(_BG_TASKS.discard)
+
+
+async def _decide_and_verdict(
+    pkt,
+    flow: tuple[str, int, str, int],
+    dst: str,
+    port: int,
+    host: str,
+    client: SidecarConsentClient,
+) -> None:
+    """Await the consent verdict for a held SYN + apply it (deferred).
+
+    ``allow`` -> learn the IP all-ports (reconnects + this connection's SYN
+    retransmits pass without re-prompting) + ``pkt.accept()`` (conntrack
+    ESTABLISHED,RELATED carries the in-flight connection); ``deny``/timeout/
+    WS-down -> ``pkt.drop`` + a temporary REJECT (tcp-reset) rule so the next
+    retransmit gets RST'd (ECONNREFUSED) instead of the kernel retransmitting
+    for ~127s. The verdict is cached against ``flow`` (the connection tuple) so
+    SYN retransmits (tcp_syn_retries) of THIS connection reuse it -- a new
+    connection (new source port) is a cache miss and re-prompts, which is what
+    makes duration=once per-connection (#2361).
+    """
+    try:
+        decision, duration = await client.request(host, port)
+    except Exception:
+        decision, duration = "deny", "once"  # timeout / loop dead -> fail-close
+    now = time.time()
+    # Bound memory under a denied-flow flood (allowed flows get learned +
+    # stop hitting NFQUEUE; only denied flows accumulate in the cache).
+    if len(_VERDICT_CACHE) > 4096:
+        _VERDICT_CACHE.clear()
+    _VERDICT_CACHE[flow] = (decision, now + VERDICT_CACHE_TTL)
+    # Populate the in-session allow-list BEFORE the iptables fork below (which
+    # yields to the executor): a SYN to a different IP of this host arriving
+    # during that yield would otherwise miss _SESSION_HOST_ALLOWS and re-prompt.
+    # Both gates (ports_for + _cb) read it (#2372). A timed allow is host-scoped
+    # too (#2434): otherwise a CDN-rotated IP of a timed-allowed host re-enters
+    # NFQUEUE, and a hold timeout there fail-closes to a deny REJECT an
+    # in-effect allow should override -- an allow that refuses.
+    ttl = _duration_ttl(duration)
+    if decision == "allow" and ttl is not None and port:
+        _add_session_host(host, port, ttl)
+    # Symmetric host-scoped memory on the deny side (#2446): a timed/forever
+    # deny is remembered by host (not just IP) so a retry -- including a
+    # CDN-rotated IP -- is denied fast without re-prompting. ``once`` (ttl
+    # None) adds nothing (per-connection). Populated before the iptables fork
+    # below so a SYN to a different IP of this host during the yield still hits
+    # _session_host_denies_ttl in _cb.
+    if decision == "deny" and ttl is not None and port:
+        _add_session_deny(host, port, ttl)
+    # Run the iptables fork (allow/reject) in the executor so it doesn't block
+    # the loop thread -- matches the DNS path's _learn_all, which also runs
+    # off the loop. The packet is retained, so verdicting after the await is
+    # safe; the rule is installed before the SYN is released.
+    loop = asyncio.get_running_loop()
+    try:
+        if decision == "allow":
+            # `once` (ttl None) -> no learn, just this connection (reconnect
+            # re-prompts); a timed duration -> learn all-ports for it.
+            if ttl is not None:
+                try:
+                    await loop.run_in_executor(None, rules.allow, dst, None, ttl)
+                except Exception:
+                    pass
+            pkt.accept()
+        else:
+            # Forge a RST directly so connect() fails fast (ECONNREFUSED) at once,
+            # independent of the conntrack/retransmit race that made the REJECT
+            # rule flaky (#2345). The REJECT rule stays as a belt-and-suspenders
+            # backstop for any retransmit the RST missed. `once` -> the short
+            # fail-close reject window; a timed duration -> that long.
+            if port:
+                try:
+                    packets._send_rst(pkt.get_payload())
+                except Exception:
+                    pass
+                reject_ttl = ttl if ttl is not None else CONSENT_REJECT_TTL
+                try:
+                    await loop.run_in_executor(
+                        None, rules.reject, dst, port, reject_ttl
+                    )
+                except Exception:
+                    pass
+            pkt.drop()
+    finally:
+        # Connection resolved (verdict cached) -> retransmits now hit the cache,
+        # not the in-flight check. Always discard, even if the verdict raised,
+        # so a stuck connection can't block the tuple forever.
+        _INFLIGHT.discard(flow)
