+++ src/klangksidecar/klangksidecar/rules.py
@@ -1,0 +1,362 @@
+"""iptables rule management + learned-IP lifecycle (#2256, #2450).
+
+allow/reject install OUTPUT ACCEPT/REJECT rules under _LOCK (atomic w.r.t.
+sweep_once); drop_for_host revokes a host's rules; the TTL sweeper
+(_async_sweeper / sweep_once) reaps expired rules. _learn_all / _record_hosts
+/ _host_for drive the learned-IP map the DNS + NFQUEUE paths share.
+"""
+
+from __future__ import annotations
+
+import asyncio
+import socket
+import subprocess
+import time
+
+from . import config
+from .config import MARK, MIN_TTL, SWEEP_INTERVAL
+from .state import _LEARNED, _LOCK, _REJECTED
+
+
+def _rule_args(ip: str, port: int | None) -> list[str]:
+    """iptables OUTPUT rule args for ``ACCEPT`` to ``ip`` (optionally scoped)."""
+    args = ["-d", ip]
+    if port is not None:
+        args += ["-p", "tcp", "--dport", str(port)]
+    args += ["-j", "ACCEPT"]
+    return args
+
+
+def _rule_exists(ip: str, port: int | None) -> bool:
+    return (
+        subprocess.run(
+            [config.IPT, "-C", "OUTPUT", *_rule_args(ip, port)],
+            capture_output=True,
+        ).returncode
+        == 0
+    )
+
+
+def _install(ip: str, port: int | None) -> None:
+    """Insert the ACCEPT rule at the top of OUTPUT if not already present."""
+    if _rule_exists(ip, port):
+        return
+    subprocess.run(
+        [config.IPT, "-I", "OUTPUT", "1", *_rule_args(ip, port)],
+        capture_output=True,
+    )
+
+
+def _remove(ip: str, port: int | None) -> None:
+    """Delete one matching ACCEPT rule; swallow failure if it's already gone."""
+    subprocess.run(
+        [config.IPT, "-D", "OUTPUT", *_rule_args(ip, port)],
+        capture_output=True,
+    )
+
+
+def allow(ip: str, port: int | None, ttl: int | float) -> None:
+    """Install (if new) the ACCEPT for ``ip[:port]`` and refresh its TTL.
+
+    ``port`` is ``None`` for an all-ports rule. The learned IP's expiry is
+    set to ``now + max(ttl, MIN_TTL)`` (a 0-TTL response must not yank the
+    rule the workspace needs to reach the IP it just resolved) and only ever
+    moves forward, so a shorter-TTL re-resolution can't prematurely expire a
+    longer-lived prior rule (#2256).
+
+    The install happens **under** :data:`_LOCK` so the kernel rule and its
+    ``_LEARNED`` record are atomic w.r.t. :func:`sweep_once`'s remove+delete
+    (also under the lock). Without that, a concurrent sweep running in another
+    executor worker could delete a rule :func:`allow` just installed while
+    ``_LEARNED`` still records it as present -- a fail-closed availability gap
+    that only self-heals on the next re-resolution (#2256 review). ``allow`` and
+    :func:`sweep_once` both run off the event loop in the default thread-pool
+    executor (see :func:`_learn_all` / :func:`_async_sweeper`), so the lock
+    genuinely serializes them; contention is negligible.
+    """
+    expire = time.time() + max(ttl, MIN_TTL)
+    with _LOCK:
+        _install(ip, port)
+        rec = _LEARNED.get(ip)
+        if rec is None:
+            _LEARNED[ip] = {
+                "expire": expire,
+                "rule_expire": expire,
+                "ports": {port},
+                "host": None,
+            }
+        else:
+            rec["expire"] = max(rec["expire"], expire)
+            # rule_expire is the ACCEPT rule's lifetime, kept SEPARATE from the
+            # host-mapping expire so a re-resolve's longer DNS TTL can't extend
+            # a consent allow's rule past its verdict (#2408). max() preserves
+            # the longest across static re-learns (#2256); for a consent allow
+            # it is just the verdict's TTL (the pre-existing rule_expire is
+            # absent -- only _record_hosts has touched the record -- and `or
+            # 0.0` coerces the None).
+            rec["rule_expire"] = max(rec.get("rule_expire") or 0.0, expire)
+            rec["ports"].add(port)
+            # ``host`` (set by _record_hosts) is preserved across re-learn.
+        # An all-ports allow (the consent path) supersedes any prior per-port
+        # denies for this IP -- otherwise the all-ports ACCEPT at the top of
+        # OUTPUT would silently shadow a lingering REJECT (the decider allowed
+        # the host, so a prior port-specific deny no longer applies).
+        if port is None:
+            for key in [k for k in _REJECTED if k[0] == ip]:
+                try:
+                    _remove_reject(*key)
+                except Exception:
+                    pass
+                del _REJECTED[key]
+
+
+def _reject_rule_args(ip: str, port: int) -> list[str]:
+    """iptables OUTPUT rule args for REJECT (tcp-reset) to ``ip:port``."""
+    return [
+        "-d",
+        ip,
+        "-p",
+        "tcp",
+        "--dport",
+        str(port),
+        "-j",
+        "REJECT",
+        "--reject-with",
+        "tcp-reset",
+    ]
+
+
+def _reject_rule_exists(ip: str, port: int) -> bool:
+    return (
+        subprocess.run(
+            [config.IPT, "-C", "OUTPUT", *_reject_rule_args(ip, port)],
+            capture_output=True,
+        ).returncode
+        == 0
+    )
+
+
+def _install_reject(ip: str, port: int) -> None:
+    """Insert the REJECT (tcp-reset) rule at the top of OUTPUT if not present."""
+    if _reject_rule_exists(ip, port):
+        return
+    subprocess.run(
+        [config.IPT, "-I", "OUTPUT", "1", *_reject_rule_args(ip, port)],
+        capture_output=True,
+    )
+
+
+def _remove_reject(ip: str, port: int) -> None:
+    """Delete the REJECT rule; swallow failure if it's already gone."""
+    subprocess.run(
+        [config.IPT, "-D", "OUTPUT", *_reject_rule_args(ip, port)],
+        capture_output=True,
+    )
+
+
+def reject(ip: str, port: int, ttl: float) -> None:
+    """Install a temporary REJECT (tcp-reset) for ``ip:port`` + set its TTL.
+
+    A denied SYN is dropped, but dropping a SYN doesn't fail ``connect()`` --
+    the kernel retransmits (tcp_syn_retries, ~127s) before timing out. The
+    REJECT rule makes the next retransmit get a RST, so ``connect()`` returns
+    ECONNREFUSED at once (eager deny). Like :func:`allow`, the install + the
+    ``_REJECTED`` record are atomic under :data:`_LOCK` w.r.t. :func:`sweep_once`.
+    """
+    expire = time.time() + ttl
+    with _LOCK:
+        _install_reject(ip, port)
+        _REJECTED[(ip, port)] = max(_REJECTED.get((ip, port), 0.0), expire)
+
+
+def drop_for_host(host: str, decision: str) -> set[str]:
+    """Drop the sidecar's rules for a host (revocation, #2339).
+
+    ``allowed``: remove the learned ACCEPT rules (+ ``_LEARNED`` records) for
+    the host's IPs (revert to default/allow-list filtering).
+    ``denied``: remove the temporary REJECT rules for the host's IPs (stop
+    force-rejecting; the host is again subject to the allow-list).
+
+    Returns the set of candidate IPs (the host's resolved IPs + the host
+    string itself, for a direct-IP connect) so the caller --
+    :meth:`SidecarConsentClient._handle_drop_rule`, on the event loop -- can
+    clear the loop-only ``_SESSION_HOST_ALLOWS``/``_VERDICT_CACHE`` state
+    (via :func:`_drop_session_hosts` / :func:`_clear_verdict_cache`). Those
+    structures are documented loop-only
+    (no lock) and must NOT be mutated here, since this function runs off the
+    loop in the executor.
+
+    Host->IP comes from ``_LEARNED[ip]["host"]`` (set by ``_record_hosts`` for
+    every resolved name, allow-listed or not), so a deny's IPs are found too;
+    the host string itself is also a candidate IP (a direct-IP connect that
+    never went through DNS, and a direct-IP allow whose ``host`` is ``None``).
+    Best-effort: a failed delete drops one rule, not the whole revoke. Sync
+    (forks iptables) -- run off the loop; under ``_LOCK`` like allow/sweep.
+
+    L3/L4 limit (co-resident hosts): the egress rules are per IP+port, so two
+    DNS names that resolve to the SAME IP share one rule and cannot be revoked
+    individually -- revoking one name removes the shared rule, affecting the
+    other too. A correct per-host revoke for co-resident hosts (CDN/S3/Cloudflare
+    fronted sites) needs L7/SNI filtering, which is a separate feature (#2352).
+    """
+    host_l = host.lower()
+    with _LOCK:
+        ips = [
+            ip
+            for ip, rec in _LEARNED.items()
+            if (rec.get("host") or "").lower() == host_l
+        ]
+        # IPs that resolved to this host, plus the host itself if it's a
+        # direct-IP connect/allow (the scan above misses a direct-IP allow,
+        # whose host record is None).
+        targets = {ip for ip in ips}
+        targets.add(host_l)
+        targets.add(host)
+        if decision == "allowed":
+            for ip in [i for i in targets if i in _LEARNED]:
+                for port in list(_LEARNED[ip]["ports"]):
+                    try:
+                        _remove(ip, port)
+                    except Exception:
+                        pass
+                del _LEARNED[ip]
+        elif decision == "denied":
+            for key in [k for k in _REJECTED if k[0] in targets]:
+                try:
+                    _remove_reject(*key)
+                except Exception:
+                    pass
+                del _REJECTED[key]
+    return targets
+
+
+def sweep_once(now: float | None = None) -> list[tuple[str, set]]:
+    """Remove ACCEPT rules whose TTL has elapsed; return ``(ip, ports)`` removed.
+
+    Two lifetimes are tracked per learned IP (#2408):
+
+    * ``rule_expire`` -- the ACCEPT rule's lifetime (a consent allow's verdict,
+      or a static re-learn's DNS TTL). When it elapses the kernel ACCEPT rule
+      is deleted but the record is KEPT while its host-mapping ``expire`` is
+      still valid, so :func:`_host_for` can still name the host for a fresh
+      consent request.
+    * ``expire`` -- the host-mapping lifetime (the DNS TTL). When it elapses
+      AND no ACCEPT rule remains, the whole record is dropped.
+
+    Records without ``rule_expire`` (host-mapping-only entries from
+    :func:`_record_hosts`, or pre-#2408 records) fall back to ``expire`` for the
+    rule sweep, preserving the old single-expiry behavior. Removal runs under
+    :data:`_LOCK` (see :func:`allow`): the rule delete and the record delete are
+    atomic, so a concurrent :func:`allow` can't re-record an IP whose kernel
+    rule was just swept. Factored out of :func:`_async_sweeper` so it is
+    unit-testable with a mocked clock and iptables (#2256).
+    """
+    if now is None:
+        now = time.time()
+    expired: list[tuple[str, set]] = []
+    with _LOCK:
+        for ip, rec in list(_LEARNED.items()):
+            # Rule sweep: the ACCEPT rule's lifetime is rule_expire when set
+            # (a consent allow, whose verdict must outlive the host-mapping's
+            # DNS TTL, #2408), else expire (static re-learn / backward compat).
+            rule_expire = rec.get("rule_expire", rec["expire"])
+            if rec["ports"] and rule_expire <= now:
+                ports = set(rec["ports"])
+                for port in ports:
+                    try:
+                        _remove(ip, port)
+                    except Exception:
+                        pass  # a transient failure drops one rule, not the sweep
+                expired.append((ip, ports))
+                rec["ports"] = set()  # rule gone; keep record for naming
+            # Record sweep: drop the host mapping once its own expire elapses
+            # and no ACCEPT rule remains.
+            if rec["expire"] <= now and not rec["ports"]:
+                del _LEARNED[ip]
+        # also sweep temporary REJECT (tcp-reset) rules for denied connections
+        for key in [k for k, exp in _REJECTED.items() if exp <= now]:
+            try:
+                _remove_reject(*key)
+            except Exception:
+                pass
+            del _REJECTED[key]
+    return expired
+
+
+async def _async_sweeper() -> None:
+    """Background task: periodically drop learned IPs past their TTL (#2256).
+
+    :func:`sweep_once` runs in the executor so its iptables ``-D`` forks don't
+    block the loop (and so it can run concurrently with :func:`_learn_all`,
+    serialized by :data:`_LOCK`).
+    """
+    while True:
+        await asyncio.sleep(SWEEP_INTERVAL)
+        try:
+            await asyncio.get_running_loop().run_in_executor(None, sweep_once)
+        except Exception:
+            pass  # a transient sweep failure defers cleanup to the next tick
+
+
+def _fmt_ports(ports: set[int | None]) -> str:
+    return "all" if None in ports else ",".join(sorted(str(p) for p in ports))
+
+
+def check_mark() -> None:
+    """Verify the proxy can set SO_MARK (needs CAP_NET_ADMIN/NET_RAW).
+
+    Without it the proxy's upstream forwards are not exempted from the nat
+    REDIRECT and loop back into itself — DNS is broken. Fail loud at startup.
+    """
+    probe = socket.socket(socket.AF_INET, socket.SOCK_DGRAM)
+    try:
+        probe.setsockopt(socket.SOL_SOCKET, socket.SO_MARK, MARK)
+    except OSError as exc:
+        raise SystemExit(
+            f"dns-proxy: cannot set SO_MARK={MARK} ({exc}); the sidecar needs "
+            "CAP_NET_ADMIN for mark-based loop-avoidance (#2264)"
+        )
+    finally:
+        probe.close()
+
+
+def _learn_all(recs: list[tuple[str, int]], ports: set[int | None]) -> None:
+    """Install the ACCEPT rule for each learned IP/port (sync; runs in the
+    executor so the iptables forks don't block the event loop)."""
+    for ip, ttl in recs:
+        for port in ports:
+            allow(ip, port, ttl)
+
+
+def _record_hosts(recs: list[tuple[str, int]], host: str) -> None:
+    """Record IP->host in _LEARNED WITHOUT installing an ACCEPT rule (#2324).
+
+    A non-allow-listed name resolves (the workspace gets the IP + can SYN) but
+    its connection is consent-gated at the SYN (NFQUEUE), so the IP must NOT be
+    allow-learned here. Recording host lets the NFQUEUE consumer name the host
+    in the consent request. Sync; runs in the executor alongside allow/sweep
+    (under _LOCK). The TTL refreshes on each resolve so a re-resolution extends
+    the window in which a SYN names the right host.
+
+    Only the host-mapping ``expire`` is touched here -- never ``rule_expire``
+    (the ACCEPT rule's lifetime, set by :func:`allow`). Keeping the two
+    separate is what lets a consent allow's rule expire at its verdict while
+    the host mapping lives for the DNS TTL, so a re-resolve's longer DNS TTL
+    can't extend an allow past its verdict (#2408).
+    """
+    now = time.time()
+    with _LOCK:
+        for ip, ttl in recs:
+            expire = now + max(ttl, MIN_TTL)
+            rec = _LEARNED.get(ip)
+            if rec is None:
+                _LEARNED[ip] = {"expire": expire, "ports": set(), "host": host}
+            else:
+                rec["expire"] = max(rec["expire"], expire)
+                rec["host"] = host  # latest name that resolved to this IP
+
+
+def _host_for(ip: str) -> str:
+    """The DNS name that resolved to ``ip``, or ``ip`` itself (direct-IP connect)."""
+    with _LOCK:
+        return _LEARNED.get(ip, {}).get("host") or ip
