+++ src/klangksidecar/pyproject.toml
@@ -1,17 +1,15 @@
 # The network sidecar, as a first-class standalone package (#2450).
 #
-# ``proxy.py`` is the FQDN egress DNS proxy + interactive consent hold
-# (#2250/#2253/#2311). It is a *copy* of ``src/containers/network/proxy.py`` —
-# the container-embedded copy is what ships in the sidecar image today and
-# must keep working unchanged; this package lets the sidecar be iterated on
-# (consent flow, future ingress/egress) without disturbing that copy and
-# without coupling the sidecar's tests to the ``klangk`` server package.
-#
-# ``proxy.py`` is a standalone script that runs as PID 1 of the sidecar
-# (``__main__``), so it stays a top-level module rather than being folded into
-# a package subdirectory. Its version constraint mirrors the sidecar image's
-# Dockerfile (``src/containers/network/Dockerfile``) so the dev/test venv
-# matches what ships.
+# The import package ``klangksidecar`` is the FQDN egress DNS proxy + the
+# interactive consent hold (#2250/#2253/#2311): config/state/allowlist/rules/
+# dns/packets/consent/nfqueue submodules + an ``app`` orchestration layer
+# (``__main__`` -> app.main). It began as a byte-for-byte copy of
+# ``src/containers/network/proxy.py`` (still shipped in the sidecar image,
+# still unit-tested by the klangk suite) and is now split into per-concern
+# modules so the sidecar can be iterated on without coupling its tests to the
+# ``klangk`` server package. Dep pins mirror the sidecar image's Dockerfile
+# (``src/containers/network/Dockerfile``) so the dev/test venv matches what
+# ships; the image installs this wheel and runs ``python3 -m klangksidecar``.
 [project]
 name = "klangksidecar"
 dynamic = ["version"]
@@ -29,7 +27,7 @@ dependencies = [
 
 # netfilterqueue is NOT a default install: it links libnetfilter_queue (a C
 # lib + dev headers, Linux-only, needs CAP_NET_ADMIN at runtime) and would
-# break installs on macOS / CI runners that lack it. proxy.py imports it
+# break installs on macOS / CI runners that lack it. nfqueue.py imports it
 # *lazily* (inside _setup_nfq_consumer) and degrades gracefully to a no-op
 # when absent — opt in here only when actually driving the NFQUEUE path.
 [project.optional-dependencies]
@@ -41,13 +39,13 @@ test = [
     "pytest-timeout>=2.3.0",
 ]
 
-# No coverage gate here, by design: proxy.py's main()/signal-handler/real-
-# socket paths only run as PID 1 of the sidecar and are exercised end-to-end
-# by the real-podman e2e (src/klangk/klangkd-tests/e2e-tests/
-# test_network_sidecar_e2e.py), not by this unit suite. Mirrors how the
-# klangk suite treats proxy.py today (it isn't part of the klangk package, so
-# it was never coverage-gated). The unit suite pins the proxy's logic
-# (spec parsing, ports_for, allow/sweep, the consent state machine).
+# No coverage gate here, by design: app.main()'s signal-handler/real-socket
+# paths only run as PID 1 of the sidecar and are exercised end-to-end by the
+# real-podman e2e (src/klangk/klangkd-tests/e2e-tests/
+# test_network_sidecar_e2e.py), not by this unit suite (mirrors how the klangk
+# suite treats the proxy today — it was never part of the klangk package, so
+# never coverage-gated). The unit suite pins the proxy's logic (spec parsing,
+# ports_for, allow/sweep, the consent state machine).
 [tool.pytest.ini_options]
 asyncio_mode = "auto"
 asyncio_default_fixture_loop_scope = "function"
@@ -58,13 +56,12 @@ filterwarnings = [
     "ignore:coroutine .* was never awaited:RuntimeWarning",
 ]
 
-# Deliberately NO ``[tool.ruff]`` override: proxy.py is a byte-for-byte copy
-# of ``src/containers/network/proxy.py``, which lives outside ``src/klangk``
-# and so is formatted at ruff's *default* line length (88). Adding
-# ``line-length = 79`` here (the klangk convention) would make ruff reformat
-# the sidecar copy and silently break the byte-for-byte invariant. Both
-# files pass ruff format/check at the default config, which is what the
-# container copy is held to.
+# Deliberately NO ``[tool.ruff]`` override: the proxy source originated as a
+# byte-for-byte copy of ``src/containers/network/proxy.py`` (which lives
+# outside ``src/klangk`` and so is formatted at ruff's *default* line length,
+# 88), and keeping the default here means the extracted submodules stay at the
+# same formatting the original was held to. All files pass ruff format/check
+# at the default config.
 
 [tool.hatch.version]
 source = "vcs"
@@ -75,11 +72,10 @@ raw-options.root = "../.."
 raw-options.tag_regex = "^v(?P<version>[vV]?\\d+(?:\\.\\d+){0,2}[^\\+]*)(?:\\+.*)?$"
 raw-options.git_describe_command = ["git", "describe", "--dirty", "--tags", "--long", "--match", "v*"]
 
-# proxy.py is a single top-level module (flat layout); ship it as the
-# importable ``proxy`` module. tests/ has no __init__.py so it is excluded
-# from the wheel automatically.
+# Ship the ``klangksidecar`` import package (the split submodules). tests/ has
+# no __init__.py so it is excluded from the wheel automatically.
 [tool.hatch.build.targets.wheel]
-packages = ["proxy.py"]
+packages = ["klangksidecar"]
 
 [build-system]
 requires = ["hatchling", "hatch-vcs"]
